+++ src/modules/uORB/Publication.cpp
@@ -37,19 +37,19 @@
  */
 
 #include "Publication.hpp"
-#include "topics/vehicle_attitude.h"
-#include "topics/vehicle_local_position.h"
-#include "topics/vehicle_global_position.h"
-#include "topics/debug_key_value.h"
-#include "topics/actuator_controls.h"
-#include "topics/vehicle_global_velocity_setpoint.h"
-#include "topics/vehicle_attitude_setpoint.h"
-#include "topics/vehicle_rates_setpoint.h"
-#include "topics/actuator_outputs.h"
-#include "topics/actuator_direct.h"
-#include "topics/encoders.h"
-#include "topics/tecs_status.h"
-#include "topics/rc_channels.h"
+#include "uORB/topics/vehicle_attitude.h"
+#include "uORB/topics/vehicle_local_position.h"
+#include "uORB/topics/vehicle_global_position.h"
+#include "uORB/topics/debug_key_value.h"
+#include "uORB/topics/actuator_controls.h"
+#include "uORB/topics/vehicle_global_velocity_setpoint.h"
+#include "uORB/topics/vehicle_attitude_setpoint.h"
+#include "uORB/topics/vehicle_rates_setpoint.h"
+#include "uORB/topics/actuator_outputs.h"
+#include "uORB/topics/actuator_direct.h"
+#include "uORB/topics/encoders.h"
+#include "uORB/topics/tecs_status.h"
+#include "uORB/topics/rc_channels.h"
 
 namespace uORB
 {
