+++ src/modules/uORB/Subscription.cpp
@@ -37,25 +37,25 @@
  */
 
 #include "Subscription.hpp"
-#include "topics/parameter_update.h"
-#include "topics/actuator_controls.h"
-#include "topics/vehicle_gps_position.h"
-#include "topics/satellite_info.h"
-#include "topics/sensor_combined.h"
-#include "topics/hil_sensor.h"
-#include "topics/vehicle_attitude.h"
-#include "topics/vehicle_global_position.h"
-#include "topics/encoders.h"
-#include "topics/position_setpoint_triplet.h"
-#include "topics/vehicle_status.h"
-#include "topics/manual_control_setpoint.h"
-#include "topics/vehicle_local_position_setpoint.h"
-#include "topics/vehicle_local_position.h"
-#include "topics/vehicle_attitude_setpoint.h"
-#include "topics/vehicle_rates_setpoint.h"
-#include "topics/rc_channels.h"
-#include "topics/vehicle_control_mode.h"
-#include "topics/actuator_armed.h"
+#include "uORB/topics/parameter_update.h"
+#include "uORB/topics/actuator_controls.h"
+#include "uORB/topics/vehicle_gps_position.h"
+#include "uORB/topics/satellite_info.h"
+#include "uORB/topics/sensor_combined.h"
+#include "uORB/topics/hil_sensor.h"
+#include "uORB/topics/vehicle_attitude.h"
+#include "uORB/topics/vehicle_global_position.h"
+#include "uORB/topics/encoders.h"
+#include "uORB/topics/position_setpoint_triplet.h"
+#include "uORB/topics/vehicle_status.h"
+#include "uORB/topics/manual_control_setpoint.h"
+#include "uORB/topics/vehicle_local_position_setpoint.h"
+#include "uORB/topics/vehicle_local_position.h"
+#include "uORB/topics/vehicle_attitude_setpoint.h"
+#include "uORB/topics/vehicle_rates_setpoint.h"
+#include "uORB/topics/rc_channels.h"
+#include "uORB/topics/vehicle_control_mode.h"
+#include "uORB/topics/actuator_armed.h"
 
 namespace uORB
 {
