+++ src/platforms/common/module.mk
@@ -1,7 +1,8 @@
 #
-# Common OS porting APIs
+# Common API used for all supported OS
 #
 
-SRCS		 = px4_getopt.c \
+SRCS		 = objects_common.cpp \
+		   px4_getopt.c \
 		   px4_log.c
