+++ src/platforms/common/objects_common.cpp
@@ -63,202 +63,202 @@ ORB_DEFINE(output_pwm, struct pwm_output_values);
 #include <drivers/drv_rc_input.h>
 ORB_DEFINE(input_rc, struct rc_input_values);
 
-#include "topics/pwm_input.h"
+#include "uORB/topics/pwm_input.h"
 ORB_DEFINE(pwm_input, struct pwm_input_s);
 
-#include "topics/vehicle_attitude.h"
+#include "uORB/topics/vehicle_attitude.h"
 ORB_DEFINE(vehicle_attitude, struct vehicle_attitude_s);
 
-#include "topics/sensor_combined.h"
+#include "uORB/topics/sensor_combined.h"
 ORB_DEFINE(sensor_combined, struct sensor_combined_s);
 
-#include "topics/hil_sensor.h"
+#include "uORB/topics/hil_sensor.h"
 ORB_DEFINE(hil_sensor, struct hil_sensor_s);
 
-#include "topics/vehicle_gps_position.h"
+#include "uORB/topics/vehicle_gps_position.h"
 ORB_DEFINE(vehicle_gps_position, struct vehicle_gps_position_s);
 
-#include "topics/vehicle_land_detected.h"
+#include "uORB/topics/vehicle_land_detected.h"
 ORB_DEFINE(vehicle_land_detected, struct vehicle_land_detected_s);
 
-#include "topics/satellite_info.h"
+#include "uORB/topics/satellite_info.h"
 ORB_DEFINE(satellite_info, struct satellite_info_s);
 
-#include "topics/home_position.h"
+#include "uORB/topics/home_position.h"
 ORB_DEFINE(home_position, struct home_position_s);
 
-#include "topics/vehicle_status.h"
+#include "uORB/topics/vehicle_status.h"
 ORB_DEFINE(vehicle_status, struct vehicle_status_s);
 
-#include "topics/vtol_vehicle_status.h"
+#include "uORB/topics/vtol_vehicle_status.h"
 ORB_DEFINE(vtol_vehicle_status, struct vtol_vehicle_status_s);
 
-#include "topics/safety.h"
+#include "uORB/topics/safety.h"
 ORB_DEFINE(safety, struct safety_s);
 
-#include "topics/battery_status.h"
+#include "uORB/topics/battery_status.h"
 ORB_DEFINE(battery_status, struct battery_status_s);
 
-#include "topics/servorail_status.h"
+#include "uORB/topics/servorail_status.h"
 ORB_DEFINE(servorail_status, struct servorail_status_s);
 
-#include "topics/system_power.h"
+#include "uORB/topics/system_power.h"
 ORB_DEFINE(system_power, struct system_power_s);
 
-#include "topics/vehicle_global_position.h"
+#include "uORB/topics/vehicle_global_position.h"
 ORB_DEFINE(vehicle_global_position, struct vehicle_global_position_s);
 
-#include "topics/vehicle_local_position.h"
+#include "uORB/topics/vehicle_local_position.h"
 ORB_DEFINE(vehicle_local_position, struct vehicle_local_position_s);
 
-#include "topics/att_pos_mocap.h"
+#include "uORB/topics/att_pos_mocap.h"
 ORB_DEFINE(att_pos_mocap, struct att_pos_mocap_s);
 
-#include "topics/vehicle_rates_setpoint.h"
+#include "uORB/topics/vehicle_rates_setpoint.h"
 ORB_DEFINE(vehicle_rates_setpoint, struct vehicle_rates_setpoint_s);
-#include "topics/mc_virtual_rates_setpoint.h"
+#include "uORB/topics/mc_virtual_rates_setpoint.h"
 ORB_DEFINE(mc_virtual_rates_setpoint, struct mc_virtual_rates_setpoint_s);
-#include "topics/fw_virtual_rates_setpoint.h"
+#include "uORB/topics/fw_virtual_rates_setpoint.h"
 ORB_DEFINE(fw_virtual_rates_setpoint, struct fw_virtual_rates_setpoint_s);
 
-#include "topics/rc_channels.h"
+#include "uORB/topics/rc_channels.h"
 ORB_DEFINE(rc_channels, struct rc_channels_s);
 
-#include "topics/vehicle_command.h"
+#include "uORB/topics/vehicle_command.h"
 ORB_DEFINE(vehicle_command, struct vehicle_command_s);
 
-#include "topics/vehicle_control_mode.h"
+#include "uORB/topics/vehicle_control_mode.h"
 ORB_DEFINE(vehicle_control_mode, struct vehicle_control_mode_s);
 
-#include "topics/vehicle_local_position_setpoint.h"
+#include "uORB/topics/vehicle_local_position_setpoint.h"
 ORB_DEFINE(vehicle_local_position_setpoint, struct vehicle_local_position_setpoint_s);
 
-#include "topics/position_setpoint_triplet.h"
+#include "uORB/topics/position_setpoint_triplet.h"
 ORB_DEFINE(position_setpoint_triplet, struct position_setpoint_triplet_s);
 
-#include "topics/vehicle_global_velocity_setpoint.h"
+#include "uORB/topics/vehicle_global_velocity_setpoint.h"
 ORB_DEFINE(vehicle_global_velocity_setpoint, struct vehicle_global_velocity_setpoint_s);
 
-#include "topics/mission.h"
+#include "uORB/topics/mission.h"
 ORB_DEFINE(offboard_mission, struct mission_s);
 ORB_DEFINE(onboard_mission, struct mission_s);
 
-#include "topics/mission_result.h"
+#include "uORB/topics/mission_result.h"
 ORB_DEFINE(mission_result, struct mission_result_s);
 
-#include "topics/geofence_result.h"
+#include "uORB/topics/geofence_result.h"
 ORB_DEFINE(geofence_result, struct geofence_result_s);
 
-#include "topics/fence.h"
+#include "uORB/topics/fence.h"
 ORB_DEFINE(fence, struct fence_s);
 
-#include "topics/fence_vertex.h"
+#include "uORB/topics/fence_vertex.h"
 ORB_DEFINE(fence_vertex, struct fence_vertex_s);
 
-#include "topics/vehicle_attitude_setpoint.h"
+#include "uORB/topics/vehicle_attitude_setpoint.h"
 ORB_DEFINE(vehicle_attitude_setpoint, struct vehicle_attitude_setpoint_s);
 ORB_DEFINE(mc_virtual_attitude_setpoint, struct vehicle_attitude_setpoint_s);
 ORB_DEFINE(fw_virtual_attitude_setpoint, struct vehicle_attitude_setpoint_s);
 
-#include "topics/manual_control_setpoint.h"
+#include "uORB/topics/manual_control_setpoint.h"
 ORB_DEFINE(manual_control_setpoint, struct manual_control_setpoint_s);
 
-#include "topics/offboard_control_mode.h"
+#include "uORB/topics/offboard_control_mode.h"
 ORB_DEFINE(offboard_control_mode, struct offboard_control_mode_s);
 
-#include "topics/optical_flow.h"
+#include "uORB/topics/optical_flow.h"
 ORB_DEFINE(optical_flow, struct optical_flow_s);
 
-#include "topics/filtered_bottom_flow.h"
+#include "uORB/topics/filtered_bottom_flow.h"
 ORB_DEFINE(filtered_bottom_flow, struct filtered_bottom_flow_s);
 
-#include "topics/airspeed.h"
+#include "uORB/topics/airspeed.h"
 ORB_DEFINE(airspeed, struct airspeed_s);
 
-#include "topics/differential_pressure.h"
+#include "uORB/topics/differential_pressure.h"
 ORB_DEFINE(differential_pressure, struct differential_pressure_s);
 
-#include "topics/subsystem_info.h"
+#include "uORB/topics/subsystem_info.h"
 ORB_DEFINE(subsystem_info, struct subsystem_info_s);
 
 /* actuator controls, as requested by controller */
-#include "topics/actuator_controls.h"
-#include "topics/actuator_controls_0.h"
+#include "uORB/topics/actuator_controls.h"
+#include "uORB/topics/actuator_controls_0.h"
 ORB_DEFINE(actuator_controls_0, struct actuator_controls_0_s);
-#include "topics/actuator_controls_1.h"
+#include "uORB/topics/actuator_controls_1.h"
 ORB_DEFINE(actuator_controls_1, struct actuator_controls_1_s);
-#include "topics/actuator_controls_2.h"
+#include "uORB/topics/actuator_controls_2.h"
 ORB_DEFINE(actuator_controls_2, struct actuator_controls_2_s);
-#include "topics/actuator_controls_3.h"
+#include "uORB/topics/actuator_controls_3.h"
 ORB_DEFINE(actuator_controls_3, struct actuator_controls_3_s);
 //Virtual control groups, used for VTOL operation
-#include "topics/actuator_controls_virtual_mc.h"
+#include "uORB/topics/actuator_controls_virtual_mc.h"
 ORB_DEFINE(actuator_controls_virtual_mc, struct actuator_controls_virtual_mc_s);
-#include "topics/actuator_controls_virtual_fw.h"
+#include "uORB/topics/actuator_controls_virtual_fw.h"
 ORB_DEFINE(actuator_controls_virtual_fw, struct actuator_controls_virtual_fw_s);
 
-#include "topics/actuator_armed.h"
+#include "uORB/topics/actuator_armed.h"
 ORB_DEFINE(actuator_armed, struct actuator_armed_s);
 
-#include "topics/actuator_outputs.h"
+#include "uORB/topics/actuator_outputs.h"
 ORB_DEFINE(actuator_outputs, struct actuator_outputs_s);
 
-#include "topics/actuator_direct.h"
+#include "uORB/topics/actuator_direct.h"
 ORB_DEFINE(actuator_direct, struct actuator_direct_s);
 
-#include "topics/multirotor_motor_limits.h"
+#include "uORB/topics/multirotor_motor_limits.h"
 ORB_DEFINE(multirotor_motor_limits, struct multirotor_motor_limits_s);
 
-#include "topics/telemetry_status.h"
+#include "uORB/topics/telemetry_status.h"
 ORB_DEFINE(telemetry_status_0, struct telemetry_status_s);
 ORB_DEFINE(telemetry_status_1, struct telemetry_status_s);
 ORB_DEFINE(telemetry_status_2, struct telemetry_status_s);
 ORB_DEFINE(telemetry_status_3, struct telemetry_status_s);
 
-#include "topics/test_motor.h"
+#include "uORB/topics/test_motor.h"
 ORB_DEFINE(test_motor, struct test_motor_s);
 
-#include "topics/debug_key_value.h"
+#include "uORB/topics/debug_key_value.h"
 ORB_DEFINE(debug_key_value, struct debug_key_value_s);
 
-#include "topics/navigation_capabilities.h"
+#include "uORB/topics/navigation_capabilities.h"
 ORB_DEFINE(navigation_capabilities, struct navigation_capabilities_s);
 
-#include "topics/esc_status.h"
+#include "uORB/topics/esc_status.h"
 ORB_DEFINE(esc_status, struct esc_status_s);
 
-#include "topics/esc_report.h"
+#include "uORB/topics/esc_report.h"
 ORB_DEFINE(esc_report, struct esc_report_s);
 
-#include "topics/encoders.h"
+#include "uORB/topics/encoders.h"
 ORB_DEFINE(encoders, struct encoders_s);
 
-#include "topics/estimator_status.h"
+#include "uORB/topics/estimator_status.h"
 ORB_DEFINE(estimator_status, struct estimator_status_s);
 
-#include "topics/vision_position_estimate.h"
+#include "uORB/topics/vision_position_estimate.h"
 ORB_DEFINE(vision_position_estimate, struct vision_position_estimate_s);
 
-#include "topics/vehicle_force_setpoint.h"
+#include "uORB/topics/vehicle_force_setpoint.h"
 ORB_DEFINE(vehicle_force_setpoint, struct vehicle_force_setpoint_s);
 
-#include "topics/tecs_status.h"
+#include "uORB/topics/tecs_status.h"
 ORB_DEFINE(tecs_status, struct tecs_status_s);
 
-#include "topics/wind_estimate.h"
+#include "uORB/topics/wind_estimate.h"
 ORB_DEFINE(wind_estimate, struct wind_estimate_s);
 
-#include "topics/rc_parameter_map.h"
+#include "uORB/topics/rc_parameter_map.h"
 ORB_DEFINE(rc_parameter_map, struct rc_parameter_map_s);
 
-#include "topics/time_offset.h"
+#include "uORB/topics/time_offset.h"
 ORB_DEFINE(time_offset, struct time_offset_s);
 
-#include "topics/mc_att_ctrl_status.h"
+#include "uORB/topics/mc_att_ctrl_status.h"
 ORB_DEFINE(mc_att_ctrl_status, struct mc_att_ctrl_status_s);
 
-#include "topics/distance_sensor.h"
+#include "uORB/topics/distance_sensor.h"
 ORB_DEFINE(distance_sensor, struct distance_sensor_s);
 
-#include "topics/camera_trigger.h"
+#include "uORB/topics/camera_trigger.h"
 ORB_DEFINE(camera_trigger, struct camera_trigger_s);
