+++ src/platforms/uORB/topics/satellite_info.h
@@ -1,6 +1,9 @@
 /****************************************************************************
  *
  *   Copyright (C) 2014 PX4 Development Team. All rights reserved.
+ *   Author: @author Thomas Gubler <thomasgubler@student.ethz.ch>
+ *           @author Julian Oes <joes@student.ethz.ch>
+ *           @author Lorenz Meier <lm@inf.ethz.ch>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -20,7 +23,7 @@
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
  * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
  * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT ,
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
  * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
  * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
  * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
@@ -32,45 +35,55 @@
  ****************************************************************************/
 
 /**
- * @file rc_parameter_map.h
- * Maps RC channels to parameters
- *
- * @author Thomas Gubler <thomasgubler@gmail.com>
+ * @file satellite_info.h
+ * Definition of the GNSS satellite info uORB topic.
  */
 
-#ifndef TOPIC_RC_PARAMETER_MAP_H
-#define TOPIC_RC_PARAMETER_MAP_H
+#ifndef TOPIC_SAT_INFO_H_
+#define TOPIC_SAT_INFO_H_
 
 #include <stdint.h>
-#include "../uORB.h"
-
-#define RC_PARAM_MAP_NCHAN 3 // This limit is also hardcoded in the enum RC_CHANNELS_FUNCTION in rc_channels.h
-#define PARAM_ID_LEN 16 // corresponds to MAVLINK_MSG_PARAM_VALUE_FIELD_PARAM_ID_LEN
+#include "uORB/uORB.h"
 
 /**
  * @addtogroup topics
  * @{
  */
 
-struct rc_parameter_map_s {
-	uint64_t timestamp;			/**< time at which the map was updated */
+/**
+ * GNSS Satellite Info.
+ */
 
-	bool valid[RC_PARAM_MAP_NCHAN];		/**< true for RC-Param channels which are mapped to a param */
+#define SAT_INFO_MAX_SATELLITES  20
 
-	int param_index[RC_PARAM_MAP_NCHAN];	/**< corresponding param index, this
-						  this field is ignored if set to -1, in this case param_id will
-						  be used*/
-	char param_id[RC_PARAM_MAP_NCHAN][PARAM_ID_LEN + 1];	/**< corresponding param id, null terminated */
-	float scale[RC_PARAM_MAP_NCHAN];	/** scale to map the RC input [-1, 1] to a parameter value */
-	float value0[RC_PARAM_MAP_NCHAN];	/** inital value around which the parameter value is changed */
-	float value_min[RC_PARAM_MAP_NCHAN];	/** minimal parameter value */
-	float value_max[RC_PARAM_MAP_NCHAN];	/** minimal parameter value */
+struct satellite_info_s {
+	uint64_t timestamp;				/**< Timestamp of satellite info */
+	uint8_t count;					/**< Number of satellites in satellite info */
+	uint8_t svid[SAT_INFO_MAX_SATELLITES]; 		/**< Space vehicle ID [1..255], see scheme below  */
+	uint8_t used[SAT_INFO_MAX_SATELLITES];		/**< 0: Satellite not used, 1: used for navigation */
+	uint8_t elevation[SAT_INFO_MAX_SATELLITES];	/**< Elevation (0: right on top of receiver, 90: on the horizon) of satellite */
+	uint8_t azimuth[SAT_INFO_MAX_SATELLITES];	/**< Direction of satellite, 0: 0 deg, 255: 360 deg. */
+	uint8_t snr[SAT_INFO_MAX_SATELLITES];		/**< dBHz, Signal to noise ratio of satellite C/N0, range 0..99, zero when not tracking this satellite. */
 };
+
+/**
+ * NAV_SVINFO space vehicle ID (svid) scheme according to u-blox protocol specs
+ * u-bloxM8-V15_ReceiverDescriptionProtocolSpec_Public_(UBX-13003221).pdf
+ *
+ * GPS		1-32
+ * SBAS		120-158
+ * Galileo	211-246
+ * BeiDou	159-163, 33-64
+ * QZSS		193-197
+ * GLONASS	65-96, 255
+ *
+ */
 
 /**
  * @}
  */
 
-ORB_DECLARE(rc_parameter_map);
+/* register this as object request broker structure */
+ORB_DECLARE(satellite_info);
 
 #endif
